Modularize unit tests and QOL changes
