+++ .gitignore
@@ -1,33 +1,12 @@
-# NI Logs
-
-log_*.dat
-.DS_Store
-# Linter config
-
-.pylintrc
-
 # IDE files and directories
-
 .idea/
 .vscode/
 .pydevproject
 .project
 
 # Python
-
 venv/
 .venv/
 
 __pycache__/
 *.py[cod]
-
-build/
-dist/
-*.egg-info/
-
-.coverage
-coverage.xml
-htmlcov
-
-*.json
-*.log
