+++ pytest.ini
@@ -1,0 +1,3 @@
+; need to tell pytest where to look for modules when importing in tests/
+[pytest]
+pythonpath = .
